Matching to Houdini build 18.6.11.

+++ src/houdini/lib/H_USD/HUSD/HUSD_Scene.C
@@ -505,7 +505,7 @@ public:
 
 private:
     UT_Map<uint64, RenderTagBucket > myBuckets;
-    UT_Map<int, uint64> myPrimBucketMap;
+    UT_Map<uint64, uint64> myPrimBucketMap;
     HUSD_Scene         &myScene;
     bool                myDirtyFlag = false;
     UT_Lock             myLock;
